+++ compliance/__init__.py
@@ -13,4 +13,4 @@
 # limitations under the License.
 """Compliance automation package."""
 
-__version__ = '1.23.1'
+__version__ = '1.24.0'